First kicad commit. Some gitignore stuff
We're trying to make a single open-source/open-hardware module that functions as booth a sequel to the telex txi and as an expander for a commercial module that we are making, Time Machine. That's why files have "TM" in them.

This commit just has interface components. None of the IC's or discrete components are there. There are no traces. We're only on day one.